move addTypes to useEventMutations hook and implement changes
