+++ src/features/events/hooks/useEventMutations.ts
@@ -1,6 +1,12 @@
-import { ZetkinEvent } from 'utils/types/zetkin';
-import { eventDeleted, eventUpdate, eventUpdated } from '../store';
+import {
+  eventDeleted,
+  eventUpdate,
+  eventUpdated,
+  typeAdd,
+  typeAdded,
+} from '../store';
 import { useApiClient, useAppDispatch } from 'core/hooks';
+import { ZetkinActivity, ZetkinEvent } from 'utils/types/zetkin';
 
 export type ZetkinEventPatchBody = Partial<
   Omit<
@@ -20,6 +26,7 @@ export type ZetkinEventPatchBody = Partial<
 export type ZetkinEventPostBody = ZetkinEventPatchBody;
 
 type useEventMutationsReturn = {
+  addType: (title: string) => void;
   cancelEvent: () => void;
   deleteEvent: () => void;
   publishEvent: () => void;
@@ -36,6 +43,15 @@ export default function useEventMutations(
 ): useEventMutationsReturn {
   const apiClient = useApiClient();
   const dispatch = useAppDispatch();
+
+  const addType = (title: string) => {
+    dispatch(typeAdd([orgId, { title }]));
+    apiClient
+      .post<ZetkinActivity>(`/api/orgs/${orgId}/activities`, { title })
+      .then((event) => {
+        dispatch(typeAdded(event));
+      });
+  };
 
   const cancelEvent = () => {
     updateEvent({
@@ -88,6 +104,7 @@ export default function useEventMutations(
   };
 
   return {
+    addType,
     cancelEvent,
     deleteEvent,
     publishEvent,
